// src/app/core/interfaces/google-account-data.ts
export interface GoogleAccountData {
  at_hash: string;
  aud: string;
  azp: string;
  email: string;
  email_verified: boolean;
  exp: number;
  family_name: string;
  given_name: string;
  iat: number;
  iss: string;
  jti: string;
  name: string;
  nbf: number;
  nonce: string;
  picture: string;
  sub: string;
}